Prevent the old Node to be taken into account
Currently, the E2E test waits for the number of Nodes to be restored
after the Machine deletion, however, the timing might be so that the
Node associated with the deleted Machine is still available and so
wrongly taken into account.

This change prevents that waiting for a legitimate new Node (i.e.
created after the CR), before continuing testing the Machine.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>