Verify Machine is actually deleted
Currently, after requesting Machine deletion, the Reconciler checks only
once if the machine was actually deleted, which is usually not the case
due to Finalizer(s) in the machine, which left the logging incomplete.

With this change, the Reconciler continues to check periodically for
Machine existence until Machine deletion or check-timeout expires.

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>