Fix wrong Condition value and reason when Node or Machine is not found
Currently if Node or Machine were not found, Succeeded condition is set
to True and the reason to RemediationFinished, which is wrong.

This change not only fixes Succeded condition to False, but introduces
also a set of new reasons:

- remediationSkippedNodeNotFound
- remediationSkippedMachineNotFound
- remediationSkippedNoControllerOwner
- remediationFailed

the latter being set when remediation fails for more generic reasons
(e.g. unability to get the proper annotations from CR or Node).

fixes https://issues.redhat.com/browse/ECOPROJECT-1490

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>