Fix Remediation Started event
Emitting remediation started event at the beginning of the reconcile is
wrong. It makes think the actual remediation has started, while only the
reconcile loop has, and in fact, MDR might detect that no actual
remediation is necessary/possible later (e.g. No Node found).

This change emits the event when the remediation has actually started,
which is after the machine deletion request succeeded. For this reason,
the previous event "Machine deletion requested" was also removed because
it became a duplicate of the RemediationStarted.

This change addresses
* https://issues.redhat.com/browse/ECOPROJECT-1833
* https://issues.redhat.com/browse/ECOPROJECT-1834
* https://issues.redhat.com/browse/ECOPROJECT-1835

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>