Enhance release workflow by adding input validation for 'create_okd_release_pr' operation and updating 'ocp_version' requirement to optional.
The inputs bundle-community-okd requieredo set ocp_version. The inpue build_and_push_images not required ocp_version but they both  share the same interface, sine it's mandatory for one of the two, it was set for both.

The input for ocp_version in the rest of medik8s operator is not required. To aligned th behavior across all operators, adding input validation task only for release to bundle-community-okd.